move builder, fix imports
