+++ apps/server/src/shared/domain/builder/index.ts
